fix patch for mcpp
Interval options were implemented as a static struct of the interval class. This did not allow including the interval header in multiple compilation units, as we obtain multiple definitions of the static member during linking. We removed the static member, since it only contains a single value for the number of digits used when printing interval-bounds and replaced the usages.

Signed-off-by: Stefan Schupp <stefan.schupp@tuwien.ac.at>